refactor(apple): move script constants to core layer
- Relocate services/apple/scripts.py to
  core/apple_script_names.py
- Eliminates circular import chain:
  metrics -> services.apple -> metrics
- Top-level import now works everywhere,
  no lazy import workaround needed

Impact: clean import graph, no linter
warnings for inline imports
